refactor: 중복되는 svg 파일 제거
- svg 파일의 크기나 색상을 변경 가능하게 수정했습니다.
- 행사 관련 api에서 이미지 핅드명이 변경되어 그에 맞게 수정했습니다.
